Big update. 2 new scripts and two templates
folderbg.sh will create images to use as background images in folders. they are based off the two new templates. this script operates in a similar way to image2icon.sh
setFolderBackground.scpt an applescript for setting the folder background on OSX. this script is called from within iconizeFolders.sh when run with the -bg option. it definately slows down the processing so only use it when you need to.